+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support Revit 2025+
 ### Updates
 - Update `ElementIdConverter` to support Revit 2024+
+### Fixes
+- Fix `ObjectJsonConverter` to support `null` value.
 
 ## [1.0.3] / 2022-10-10
 ### Features